+++ tests/test_show.py
@@ -7,12 +7,12 @@ from showstats.showstats import show_stats
 def test_show(sample_df, capsys):
     show_stats(sample_df)
     captured = capsys.readouterr()
-    assert "Col (N=500)" in captured.out
+    assert "(N=500)" in captured.out
     assert "float_mean_2" in captured.out
     assert "float_min_-7" in captured.out
     show_stats(sample_df, "cat")
     captured = capsys.readouterr()
-    assert "Col (N=500)" in captured.out
+    assert "(N=500)" in captured.out
     assert "float_mean_2" not in captured.out
     assert "float_min_-7" not in captured.out
     assert "str_col" in captured.out
@@ -20,7 +20,7 @@ def test_show(sample_df, capsys):
     assert "categorical_col" in captured.out
     show_stats(sample_df, "time")
     captured = capsys.readouterr()
-    assert "Col (N=500)" in captured.out
+    assert "(N=500)" in captured.out
     assert "float_mean_2" not in captured.out
     assert "float_min_-7" not in captured.out
     assert "date_col" in captured.out
@@ -33,7 +33,7 @@ def test_show(sample_df, capsys):
 def test_show_subsets(sample_df, capsys):
     show_stats(sample_df.select("U", "int_col"))
     captured = capsys.readouterr()
-    assert "Col (N=500)" in captured.out
+    assert "(N=500)" in captured.out
     assert "int_col" in captured.out
 
     show_stats(sample_df.select("categorical_col"))